Move tests from test_redis_asyncio to main test_fakeredis
